fix bug where the `resources` field in the Kyverno test file was not considered
The `resource` field has been deprecated in favor of the `resources` field.
Previously, when running a command like:

    kyverno test . --test-case-selector "resource=nginx-deploy-pass"

the flag only considered the `resource` field in the test file, ignoring the `resources` field if present.

This commit updates the filtering logic to apply the flag to the elements of the `resources` field if it is present.

This commit does the following:
- If both `resource` and `resources` fields are present, only `resources` is considered.
- If only the `resource` field is present, it is used.

Signed-off-by: Chandan-DK <chandan.dk@nirmata.com>